refactor: emotion @shared 폴더 수정
