+++ apps/web/src/features/select-purchase-emotion/index.ts
@@ -1,4 +1,1 @@
-export { emotions } from './model';
-export type { Emotion, EmotionValue } from './model';
-
 export { EmotionSelector, EmotionDescription } from './ui';
