--- apps/web/src/features/select-purchase-emotion/model/index.ts
@@ -1,3 +1,0 @@
-export type { Emotion, EmotionValue } from './types';
-
-export { emotions } from './emotion';
