+++ apps/web/src/shared/types/index.ts
@@ -1,0 +1,1 @@
+export type { Emotion, EmotionValue } from './emotion.types';
